refactor: move node back to aot
