+++ crates/aot/src/cli.rs
@@ -29,6 +29,10 @@ pub struct Cli<N: Network> {
     /// The verbosity level of the logs.
     #[arg(long, default_value_t = 4)]
     pub verbosity: u8,
+    #[cfg(feature = "node")]
+    /// The optional loki url to send logs to.
+    #[arg(long)]
+    pub loki: Option<reqwest::Url>,
 
     #[clap(subcommand)]
     pub command: Command<N>,
@@ -48,6 +52,8 @@ pub enum Command<N: Network> {
     Man(snops_common::mangen::Mangen),
     #[cfg(feature = "clipages")]
     Md(snops_common::clipages::Clipages),
+    #[cfg(feature = "node")]
+    Run(crate::runner::Runner<N>),
 }
 
 pub trait Flushable {
@@ -80,52 +86,52 @@ pub fn make_env_filter(verbosity: u8) -> EnvFilter {
     };
 
     // Filter out undesirable logs. (unfortunately EnvFilter cannot be cloned)
-    tracing_subscriber::EnvFilter::builder()
+    let filter = tracing_subscriber::EnvFilter::builder()
         .with_env_var("AOT_LOG")
         .with_default_directive(level.into())
         .from_env_lossy()
         .add_directive("mio=off".parse().unwrap())
         .add_directive("tarpc=off".parse().unwrap())
         .add_directive("tokio_util=off".parse().unwrap())
-        // .add_directive("tokio_tungstenite=off".parse().unwrap())
-        // .add_directive("tracing_tungstenite=off".parse().unwrap())
-        // .add_directive("tungstenite=off".parse().unwrap())
+        .add_directive("tokio_tungstenite=off".parse().unwrap())
+        .add_directive("tracing_tungstenite=off".parse().unwrap())
+        .add_directive("tungstenite=off".parse().unwrap())
         .add_directive("hyper=off".parse().unwrap())
         .add_directive("reqwest=off".parse().unwrap())
         .add_directive("want=off".parse().unwrap())
-    // .add_directive("warp=off".parse().unwrap());
+        .add_directive("warp=off".parse().unwrap());
 
-    // let filter = if verbosity >= 2 {
-    //     filter.add_directive("snarkos_node_sync=trace".parse().unwrap())
-    // } else {
-    //     filter.add_directive("snarkos_node_sync=debug".parse().unwrap())
-    // };
+    let filter = if verbosity >= 2 {
+        filter.add_directive("snarkos_node_sync=trace".parse().unwrap())
+    } else {
+        filter.add_directive("snarkos_node_sync=debug".parse().unwrap())
+    };
 
-    // let filter = if verbosity >= 3 {
-    //     filter
-    //         .add_directive("snarkos_node_bft=trace".parse().unwrap())
-    //         .add_directive("snarkos_node_bft::gateway=debug".parse().
-    // unwrap()) } else {
-    //     filter.add_directive("snarkos_node_bft=debug".parse().unwrap())
-    // };
+    let filter = if verbosity >= 3 {
+        filter
+            .add_directive("snarkos_node_bft=trace".parse().unwrap())
+            .add_directive("snarkos_node_bft::gateway=debug".parse().unwrap())
+    } else {
+        filter.add_directive("snarkos_node_bft=debug".parse().unwrap())
+    };
 
-    // let filter = if verbosity >= 4 {
-    //     filter.add_directive("snarkos_node_bft::gateway=trace".parse().
-    // unwrap()) } else {
-    //     filter.add_directive("snarkos_node_bft::gateway=debug".parse().
-    // unwrap()) };
+    let filter = if verbosity >= 4 {
+        filter.add_directive("snarkos_node_bft::gateway=trace".parse().unwrap())
+    } else {
+        filter.add_directive("snarkos_node_bft::gateway=debug".parse().unwrap())
+    };
 
-    // let filter = if verbosity >= 5 {
-    //     filter.add_directive("snarkos_node_router=trace".parse().unwrap())
-    // } else {
-    //     filter.add_directive("snarkos_node_router=debug".parse().unwrap())
-    // };
+    let filter = if verbosity >= 5 {
+        filter.add_directive("snarkos_node_router=trace".parse().unwrap())
+    } else {
+        filter.add_directive("snarkos_node_router=debug".parse().unwrap())
+    };
 
-    // if verbosity >= 6 {
-    //     filter.add_directive("snarkos_node_tcp=trace".parse().unwrap())
-    // } else {
-    //     filter.add_directive("snarkos_node_tcp=off".parse().unwrap())
-    // }
+    if verbosity >= 6 {
+        filter.add_directive("snarkos_node_tcp=trace".parse().unwrap())
+    } else {
+        filter.add_directive("snarkos_node_tcp=off".parse().unwrap())
+    }
 }
 
 impl<N: Network> Cli<N> {
@@ -197,8 +203,52 @@ impl<N: Network> Cli<N> {
             )
         };
 
-        non_blocking_appender!(stderr = (io::stderr()));
-        layers.push(tracing_subscriber::fmt::layer().with_writer(stderr).boxed());
+        // Initialize tracing.
+        // Add layer using LogWriter for stdout / terminal
+        match self.command {
+            #[cfg(feature = "node")]
+            Command::Run(_) => {
+                use crossterm::tty::IsTty;
+                non_blocking_appender!(stdout = (io::stdout()));
+                layers.push(
+                    tracing_subscriber::fmt::layer()
+                        .with_ansi(io::stdout().is_tty())
+                        .with_thread_ids(true)
+                        .with_writer(stdout)
+                        .boxed(),
+                );
+            }
+            _ => {
+                non_blocking_appender!(stderr = (io::stderr()));
+                layers.push(tracing_subscriber::fmt::layer().with_writer(stderr).boxed());
+            }
+        }
+
+        #[cfg(feature = "node")]
+        if let Some(loki) = &self.loki {
+            let mut builder = tracing_loki::builder();
+
+            let env_var = std::env::var("SNOPS_LOKI_LABELS").ok();
+            let fields = match &env_var {
+                Some(var) => var
+                    .split(',')
+                    .map(|item| item.split_once('=').unwrap_or((item, "")))
+                    .collect(),
+                None => vec![],
+            };
+
+            for (key, value) in fields {
+                builder = builder.label(key, value).expect("bad loki label");
+            }
+
+            let (layer, task) = builder.build_url(loki.to_owned()).expect("bad loki url");
+            std::thread::spawn(|| {
+                let rt = tokio::runtime::Runtime::new().unwrap();
+                let handle = rt.spawn(task);
+                rt.block_on(handle).unwrap();
+            });
+            layers.push(layer.boxed());
+        };
 
         tracing_subscriber::registry()
             .with(env_filter)
@@ -224,6 +274,8 @@ impl<N: Network> Cli<N> {
             ),
             #[cfg(feature = "clipages")]
             Command::Md(clipages) => clipages.run::<Cli<N>>(env!("CARGO_PKG_NAME")),
+            #[cfg(feature = "node")]
+            Command::Run(command) => command.parse(log_level_handler),
         }
     }
 }
