--- crates/node/Cargo.toml
@@ -1,56 +1,0 @@
-[package]
-name = "snops-node"
-version = "0.1.0"
-edition = "2021"
-license = "MIT"
-description = "A snarkops node which wraps the snarkos node runner and adds extra functionality"
-
-[features]
-default = []
-cuda = ["snarkvm/cuda"]
-flame = ["dep:tracing-flame"]
-docpages = ["clipages", "mangen"]
-clipages = ["snops-common/clipages"]
-mangen = ["snops-common/mangen"]
-
-[dependencies]
-aleo-std.workspace = true
-anyhow.workspace = true
-bech32.workspace = true
-bincode.workspace = true
-clap = { workspace = true, features = ["env"] }
-colored.workspace = true
-crossterm.workspace = true
-futures-util.workspace = true
-http.workspace = true
-indexmap.workspace = true
-metrics-exporter-prometheus.workspace = true
-num_cpus.workspace = true
-rand.workspace = true
-rand_chacha.workspace = true
-rayon.workspace = true
-reqwest = { workspace = true, features = ["blocking", "json"] }
-serde_json.workspace = true
-snarkos-account.workspace = true
-snarkos-node.workspace = true
-snarkos-node-metrics.workspace = true
-snarkvm = { workspace = true, features = ["ledger"] }
-snops-checkpoint = { workspace = true, features = ["write"] }
-snops-common.workspace = true
-tarpc.workspace = true
-tokio.workspace = true
-tokio-tungstenite.workspace = true
-tracing.workspace = true
-tracing-appender.workspace = true
-tracing-flame = { workspace = true, optional = true }
-tracing-loki = "0.2.4"
-tracing-subscriber.workspace = true
-# rocksdb = { workspace = true, features = ["lz4"] }
-
-
-[target.'cfg(all(target_os = "linux", target_arch = "x86_64"))'.dependencies]
-tikv-jemallocator = "0.6"
-
-# [dependencies.metrics]
-# package = "snarkos-node-metrics"
-# path = "../../../snarkos/node/metrics"
