+++ scripts/agent_watch.sh
@@ -1,7 +1,6 @@
 #!/usr/bin/env bash
 
 cargo watch -x 'build --profile release-big -p snops-agent' \
-  -w ./crates/snops-agent \
-  -w ./crates/snops-common \
-  -w ./crates/snops-checkpoint \
-  -w ./crates/snops-node
+  -w ./crates/agent \
+  -w ./crates/common \
+  -w ./crates/checkpoint
