+++ scripts/control_plane_watch.sh
@@ -2,5 +2,5 @@
 
 cargo watch -x 'run -p snops' \
   -w ./crates/snops \
-  -w ./crates/snops-common \
-  -w ./crates/snops-checkpoint
+  -w ./crates/common \
+  -w ./crates/checkpoint
